move duplicate modal to saved objects management for reuse
