--- src/plugins/saved_objects/public/duplicate_modal/index.ts
@@ -1,13 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- *
- * Any modifications Copyright OpenSearch Contributors. See
- * GitHub history for details.
- */
-
-export { showDuplicateModal } from './show_duplicate_modal';
-export { SavedObjectsDuplicateModal, DuplicateMode } from './duplicate_modal';
