+++ src/plugins/saved_objects/public/index.ts
@@ -38,7 +38,6 @@ export {
   SaveResult,
   showSaveModal,
 } from './save_modal';
-export { showDuplicateModal, SavedObjectsDuplicateModal, DuplicateMode } from './duplicate_modal';
 export { getSavedObjectFinder, SavedObjectFinderUi, SavedObjectMetaData } from './finder';
 export {
   SavedObjectLoader,
