+++ src/plugins/saved_objects_management/public/management_section/objects_table/components/index.ts
@@ -32,3 +32,5 @@ export { Header } from './header';
 export { Table } from './table';
 export { Flyout } from './flyout';
 export { Relationships } from './relationships';
+export { showDuplicateModal } from './show_duplicate_modal';
+export { SavedObjectsDuplicateModal, DuplicateMode } from './duplicate_modal';
